server/pyproject.toml
[project]
name = "server"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "dj-database-url>=3.0.1",
    "django>=5.2.5",
    "djangorestframework>=3.16.1",
    "djangorestframework-simplejwt>=5.3.1",
    "django-cors-headers>=4.4.0",
    "gunicorn>=23.0.0",
    "psycopg>=3.2.9",
    "psycopg-binary>=3.2.9",
    "python-decouple>=3.8",
]
